+++ src/lua/CommunityBalanceMod/Scripts/GUIBetaBalanceChangelogData.lua
@@ -32,19 +32,19 @@ gChangelogData =
   - Are able to use ink, echo and healwave without a hivetype
   - Normal hitbox
   - Increased HP to 300%
-  - Reduced Speed by 50%
+  - Reduced Speed by 25%
   - Cannot be echo'd
   - Restricted to max 1 of each kind
   - Access to 1 new ability each (3 tres, 10 sec cooldown)
     - Umbra(Crag): Cast Umbra on all Structures in healrange
-    - Hallucinations(Shade): Create 5 fake moveable Structures for 30 seconds
-    - Stormcloud(Shift): Increase aliens movement speed by 30% for 5 seconds
+    - Hallucinations(Shade): Create 5 fake moveable Structures for 120 seconds
+    - Stormcloud(Shift): Increase aliens movement speed by 30% for 10 seconds
 	- Frenzy(Whip): 2x attack speed and 4x movement speed for 7.5 seconds
 
 »»» Whips
   - Fully matured whips attack without infestation
   - Added Bile Splash Ability
-  - Cast 3 Bile Bombs over itself (3 tres, 10 second cooldown). 
+	- Cast 3 Bile Bombs over itself (3 tres, 10 second cooldown). 
 
 »»» Shift
   - Reduced energy regenerate rate by 66%